test(logic): split move validator tests under 400 lines for #2288
Replace monolithic move_validator_test.dart with part1–part3 tests and
shared move_validator_test_support.dart. No behavior change; all 14
cases preserved.
